fix: sanitize label_file content and URL-encode the podup.config-files label (#1361)
The label_file reader split each line on '=' and inserted the (key, value)
pair into SpecGenerator.labels with no control-character or size check.
Podman JSON-encodes the value on the wire, so wire injection was never the
concern — the concern is downstream consumers that re-parse the label value
(logs, ls/inspect UIs, label-based filters), where a control character in
either side lets one entry break out of its context. The 16 MiB read cap on
the file also left the resulting HashMap size, key length, and value length
unconstrained.

This change adds sanitize_kv_pair, which validates a single key/value pair:

* key: non-empty, no ASCII control characters, length <= 253 bytes (Podman's
  per-label cap);
* value: no ASCII control characters, length <= 4 KiB;
* map: capped at 64 distinct entries, with overwrites of an existing key
  still allowed at the cap (the cap bounds distinct keys, not insertions).

build_label_file_labels now returns Result<HashMap<String, String>> so a
rejection surfaces as an Unsupported error naming the file and line, rather
than a silently-truncated map.

The podup.config-files label joins compose-file paths with ',' and so could
visually merge with the next entry when a path itself contained a ','. Each
path is now passed through encode_path_for_label which replaces ',' with
'%2C', making the round-trip through any downstream ','-split unambiguous.

Tests cover all four sanitize_kv_pair rejection cases plus the pass case,
the integration of build_label_file_labels (cap at exactly 64, over-cap
rejection with a clear file+line error, control-char rejection), and the
end-to-end encode_path_for_label behaviour on a real Podman round-trip.
