+++ tests/engine_integration/health_targeting.rs
@@ -419,204 +419,4 @@ fn build_with_env_arg() {
 	});
 }
 
-// ---------------------------------------------------------------------------
-// label_file: load labels from file (container.rs L73-74)
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn label_file_labels_applied() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let dir = tempfile::tempdir().unwrap();
-	fs::write(
-		dir.path().join("svc.labels"),
-		b"com.example.role=web\ncom.example.env=test\n",
-	)
-	.unwrap();
-	let proj = proj("lfl");
-	let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    label_file: svc.labels\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	// Both labels from the file have to reach the container. Labels are container
-	// config, which the library does not return, so read them out of band — the
-	// same way the sibling tests read annotations. Reading both catches a loader
-	// that stops after the first line.
-	let inspect = |key: &str| {
-		String::from_utf8_lossy(
-			&std::process::Command::new("podman")
-				.args([
-					"inspect",
-					&format!("{proj}-web-1"),
-					"--format",
-					&format!("{{{{index .Config.Labels \"{key}\"}}}}"),
-				])
-				.output()
-				.expect("podman inspect")
-				.stdout,
-		)
-		.trim()
-		.to_string()
-	};
-	let role = inspect("com.example.role");
-	let env = inspect("com.example.env");
-	engine.down(&file).await.unwrap();
-
-	assert_eq!(
-		role, "web",
-		"the first label_file entry did not reach the container"
-	);
-	assert_eq!(
-		env, "test",
-		"the second label_file entry did not reach the container"
-	);
-}
-
-// ---------------------------------------------------------------------------
-// optional dep not in file (lifecycle.rs L45, L70)
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn optional_dep_not_in_file() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("odf");
-	let engine = Engine::new(client, proj.clone());
-	// ghost_db not in services + required:false → resolve_order skips it,
-	// target_set pushes it (file.services.get → None → L45),
-	// dep-wait loop hits None => continue (L70)
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    depends_on:\n      ghost_db:\n        condition: service_started\n        required: false\n",
-	)
-	.unwrap();
-
-	engine
-		.up_with_options(&file, false, &[], &["web".to_string()], false, false, false)
-		.await
-		.unwrap();
-	let mut names = engine
-		.test_project_container_names()
-		.await
-		.unwrap_or_default();
-	names.sort();
-	engine.down(&file).await.unwrap();
-
-	// web has to actually come up. An optional dependency that is not in the file
-	// must be skipped, not turned into a container and not made to block the
-	// service that named it.
-	assert_eq!(
-		names,
-		vec![format!("{proj}-web-1")],
-		"an optional dependency missing from the file did not resolve to exactly web"
-	);
-}
-
-// ---------------------------------------------------------------------------
-// duplicate target_services triggers continue in target_set (lifecycle.rs L37)
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn target_services_duplicate_entry() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("tde");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	// Passing "web" twice causes it to be pushed to the target_set stack twice;
-	// the second pop finds "web" already in the set → !set.insert → continue (L37).
-	engine
-		.up_with_options(
-			&file,
-			false,
-			&[],
-			&["web".to_string(), "web".to_string()],
-			false,
-			false,
-			false,
-		)
-		.await
-		.unwrap();
-	let names = engine
-		.test_project_container_names()
-		.await
-		.unwrap_or_default();
-	engine.down(&file).await.unwrap();
-
-	// One container, not two. The deduplication is the whole point of the test,
-	// and a second pass over the same service would have shown up here rather
-	// than in the return value.
-	assert_eq!(
-		names,
-		vec![format!("{proj}-web-1")],
-		"naming the same service twice did not resolve to a single container"
-	);
-}
-
-// ---------------------------------------------------------------------------
-// service_healthy honors a healthcheck baked into the image, even when the
-// compose service declares none (health.rs / lifecycle.rs gate). The db image
-// carries its own HEALTHCHECK; web depends on it with condition: service_healthy
-// and no compose healthcheck. `up` must wait for the inherited check to report
-// healthy and then succeed.
-// ---------------------------------------------------------------------------
-
-#[test]
-fn service_healthy_image_inherited_healthcheck() {
-	let rt = tokio::runtime::Runtime::new().unwrap();
-	rt.block_on(async {
-		let client = match podman().await {
-			Some(d) => d,
-			None => return,
-		};
-		let dir = tempfile::tempdir().unwrap();
-		let proj = proj("ihc");
-		let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-		let image_tag = format!("podup-test-ihc-{}:latest", std::process::id());
-		let yaml = format!(
-			"services:\n  db:\n    build:\n      context: .\n      dockerfile_inline: |\n        FROM alpine:latest\n        HEALTHCHECK --interval=1s --timeout=2s --retries=3 CMD true\n    image: {image_tag}\n    command: [\"sleep\", \"infinity\"]\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    depends_on:\n      db:\n        condition: service_healthy\n"
-		);
-		let file = parse_str(&yaml).unwrap();
-
-		// The point is that db has NO healthcheck in the compose file — it inherits
-		// one from the image's HEALTHCHECK line, and web's service_healthy gate has
-		// to honour it. Two things can go wrong invisibly: the inherited check is
-		// ignored (web starts immediately, gate skipped) or it is never satisfied
-		// (up hangs to the deadline). Bounding the time catches the second; the
-		// containers being up catches the first collapsing into a silent no-op.
-		let started = std::time::Instant::now();
-		engine.up(&file).await.unwrap();
-		let elapsed = started.elapsed();
-		let mut names = engine
-			.test_project_container_names()
-			.await
-			.unwrap_or_default();
-		names.sort();
-		engine.down(&file).await.unwrap();
-		let _ = std::process::Command::new("podman")
-			.args(["rmi", "-f", &image_tag])
-			.status();
-
-		assert!(
-			elapsed < std::time::Duration::from_secs(60),
-			"up took {elapsed:?} — an image-inherited healthcheck must be reachable, not waited on to the deadline"
-		);
-		assert_eq!(
-			names,
-			vec![format!("{proj}-db-1"), format!("{proj}-web-1")],
-			"web did not come up behind a healthcheck inherited from the image"
-		);
-	});
-}
+// label_file: load tests moved to label_file_safety.rs (#1361).
